opentracing: handle case where parent span is nil
If there was no parent span in the `ctx` received, the span that's
given out will have a `s.parent == nil`. The `wrappedConn` tries
to use the span nevertheless and this creates a panic by nil pointer
dereference.

Since a nil parent is a valid operation,  the `span` with `parent ==
nil` should be a no-op for methods on it, and creating a new span from
it should make a root span - one that is not a child of another span
context.