Adding support for useful / common methods
A variety of methods are needed and used when dealing with resultsets from DBs.
Without these methods a lot of the functionality goes away and we do not want that.

Because of the above, code that uses this library will not have full support of the underlying functionality and this can easily become a problem.

This commit adds support for some common methods that are supported by the Rows interface, even though they are not part of the base interface.
Nevertheless, having them is important to increase coverage of different use cases.